+++ Material.Blazor/Components/Carousel/MBCarousel.razor.cs
@@ -40,11 +40,8 @@ public partial class MBCarousel<TItem> : InputComponent<int>
     private CancellationTokenSource TokenSource { get; set; } = new();
     internal InternalCarouselPanel<TItem> ICP { get; set; }
     private int ItemIndex { get; set; } = 0;
-    private bool IsRTL { get; set; } = false;
     private int RadioItemIndex { get; set; }
     private List<MBSelectElement<int>> RadioElements { get; set; }
-    private string PreviousIcon => IsRTL ? "navigate_next" : "navigate_before";
-    private string NextIcon => IsRTL ? "navigate_before" : "navigate_next";
 
 
 
@@ -78,8 +75,6 @@ public partial class MBCarousel<TItem> : InputComponent<int>
         {
             RadioElements.Add(new() { SelectedValue = i, Label = $"Image {i}" });
         }
-
-        IsRTL = await IsElementRTL(ElementReference);
     }